AzureFunctions/pyproject.toml
[tool.poetry]
name = "portuguese_politics"
version = "0.1.0"
description = "Portuguese Politics Daily Updater."
authors = ["Luis Silva <luisiul359@gmail.com>"]
packages = [ ]

[tool.poetry.dependencies]
python = "~3.9"
requests = "~2.25.1"
tqdm = "~4.62.3"
azure-storage-blob = "^12.12.0"
azure-functions = "^1.1.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
